backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kubernetes==28.1.0
python-multipart==0.0.6
pydantic==2.5.2
PyYAML==6.0.1
PyJWT